+++ update-vim-linux.sh
@@ -5,5 +5,8 @@ cd /root/.vim/bundle
 cd /root/
 tar zcvf vim.tgz .vim*
 cd ${BASEDIR}
-unlink output/vim.tgz
+if [  -f output/vim.tgz ]
+then
+    unlink output/vim.tgz
+fi
 mv vim.tgz output/
